+++ CHANGELOG.markdown
@@ -23,6 +23,12 @@ Version 1.3
 * Error message into the console in case target image properties were deemed insufficient.
 * New `rowlock` and `framelock` options allow mouse interaction lock of axis in multirow setups (GH-191).
 * New option `orientable` enables gyroscope support on compatible devices (GH-92).
+* Data key `dimensions` replaced with `width` and `height` keys.
+* URLs containing special markup (`@W`, `@H` or `@T`) will get it substituted with actualy
+  values (`width` and `height` data values or timestamp).
+* New `$.reel.substitutes` object for custom substitution methods.
+* New `responsive` option adds capability to scale in responsive page layouts (GH-222).
+
 
 Version 1.2.1
 -------------